+++ prettier.config.js
@@ -1,6 +1,0 @@
-const openWcConfig = require('@open-wc/prettier-config');
-
-// Customize the @open-wc prettier configuration a bit…
-openWcConfig.printWidth = 140;
-
-module.exports = openWcConfig;